# infra/contracts/s9-openapi.yaml
openapi: 3.1.0
info:
  title: worldview-gateway
  version: 2025.6.0
paths:
...
  /v1/alerts/{alert_id}/ack:
    delete:
      summary: Acknowledge Alert
      description: 'Proxy DELETE /api/v1/alerts/{alert_id}/ack → S10 Alert service.


        Requires authentication. Forwards X-Internal-JWT so S10 can verify the user

        owns the alert before acknowledging it (PRD-0025 §T-D-1-10).'
      operationId: acknowledge_alert_v1_alerts__alert_id__ack_delete
      parameters:
      - name: alert_id
        in: path
        required: true
        schema:
          type: string
          title: Alert Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Acknowledge Alert V1 Alerts  Alert Id  Ack Delete
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/alerts/{alert_id}/acknowledge:
    patch:
      summary: Acknowledge Alert Entity
      description: 'Proxy PATCH /api/v1/alerts/{alert_id}/acknowledge → S10.


        Forwards the (optional) JSON body and X-Internal-JWT. ``Cache-Control:

        no-store`` prevents any intermediary from caching the mutation response.'
      operationId: acknowledge_alert_entity_v1_alerts__alert_id__acknowledge_patch
      parameters:
      - name: alert_id
        in: path
        required: true
        schema:
          type: string
          title: Alert Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/alerts/{alert_id}/snooze:
    patch:
      summary: Snooze Alert Entity
      description: Proxy PATCH /api/v1/alerts/{alert_id}/snooze → S10.
      operationId: snooze_alert_entity_v1_alerts__alert_id__snooze_patch
      parameters:
      - name: alert_id
        in: path
        required: true
        schema:
          type: string
          title: Alert Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...